Interface config added (pretty much defaults)
Wg0, Wg1[0-9] are the default generated names in WireGuard (you can rename it to whatever you want). 

I did not included the public and private keys (it does not matter if you upload your config with keys btw) but I did not because I might change the keys next year anyway.